enterprise/lifecycle: move signal handlers to singals.py
